feat : Moved register command flow within main.go (#45)
* Feat : Moved register command flow within main.go

* Refactor : Updated vars & functions name

* Update commands/register/register.go


* Chore : Minor change

* Chore : Minor change

---------
