+++ Makefile
@@ -37,11 +37,6 @@ clean:
 	@rm -rf coverage.out
 	@rm -rf coverage.html
 
-register:
-	@echo "Registering commands..."
-	@go run commands/main/register.go
-	@echo "Registration complete."
-
 test-cover:
 ifeq ($(FORCE),1)
 	@echo "Force flag detected. Cleaning before tests..."
