+++ commands/register/register_test.go
@@ -1,4 +1,4 @@
-package main
+package register
 
 import (
 	"testing"
@@ -11,23 +11,23 @@ import (
 func TestInit(t *testing.T) {
 
 	t.Run("should panic when SetupConnection returns an error", func(t *testing.T) {
-		originalNewDiscord := NewDiscordSession
-		defer func() { NewDiscordSession = originalNewDiscord }()
-		NewDiscordSession = func(token string) (s *discordgo.Session, err error) {
+		originalNewDiscord := NewDiscord
+		defer func() { NewDiscord = originalNewDiscord }()
+		NewDiscord = func(token string) (s *discordgo.Session, err error) {
 			return nil, assert.AnError
 		}
-		assert.Panics(t, main)
+		assert.Panics(t, SetupRegister)
 	})
 	t.Run("should call AddHandler method of session if SetupConnection succeeds", func(t *testing.T) {
-		originalNewDiscord := NewDiscordSession
-		defer func() { NewDiscordSession = originalNewDiscord }()
-		NewDiscordSession = func(token string) (s *discordgo.Session, err error) {
+		originalNewDiscord := NewDiscord
+		defer func() { NewDiscord = originalNewDiscord }()
+		NewDiscord = func(token string) (s *discordgo.Session, err error) {
 			mockSession := &discordgo.Session{
 				State: &discordgo.State{},
 			}
 			return mockSession, nil
 		}
-		assert.Panics(t, main)
+		assert.Panics(t, SetupRegister)
 	})
 }
